# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AkriPharmacy API
  version: 24.0.0
  description: >-
    API backend para AkriPharmacy con inventario farmacéutico multisede, cadena de frío,
    dispensación, solicitudes de compra entre sedes, gestión de perfiles/usabilidades,
    trazabilidad terminada y centro operativo integrado.
servers:
  - url: http://localhost:8080/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /multisite/purchase-requests/{id}:
    get:
      summary: Obtener detalle de solicitud de compra entre sedes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
    patch:
      summary: Actualizar estado de una solicitud desde sede central
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
...